Bind all Alt names, and service requests using the first open factory.
Given a dtab: /svc => /#/namer/one | /#/namer/two

If the first alternate closes, then allow retries on the second alternate.
That may happen when the first service short-circuits after yielding a failure.

This is a proof of concept; there should be a feature toggle, and
at least one unit test fails.